refactor tasks page
